Fixed variable errors, added warning for v_o=a_o=0
In previous push, there were some silly syntax errors, like variables
being used before they were defined. Those have been fixed. Furthermore,
I forgot to mention in the previous push, that the heatlab_demo file has
been changed to use the Latch struct so that users may understand what
the Latch struct is.

Also, if the user calls solve_model and the initial velocity and
acceleration are both 0, then the code throws a warning. While it may be
physically possible for the latch to start with 0 velocity and
acceleration, that is not our focus. For the time being, we are assuming
that the latch will be pulled back with a force that is time indepenent,
in which case having 0 initial velocity and acceleration would mean
that the system never unlatches. So, the solve_model code has been
modified such that if both initial velocity and initial acceleration
are zero, the code gives a warning and returns zeros.