Modified solve_model to save t,y,ydot to csv file
Currently, the program creates a directory called "/output", where the
csv files are saved. Currently, the path of each csv file is
"/output/output--<date and time>". This should be easy to change. The
CSV  files currently have 11 headers in the top row, corresponding to
variables we're looking for. Then, each row contains physical values for
those variables at a given time. Currently, we just have t, y, and ydot.
In the future, we'll want to include all the other variables, which will
require us to calculate them (shouldn't be too difficult), and then edit
the matlab code to write the additional information to the CSV file.
Writing the other stuff shouldn't be too difficult. Either we add the
other information to the 'sol' matrix, and write that larger matrix to
the csv, or we figure out a way to make the writematrix function append
the numbers to the file in such a way that doesn't overwrite the
data from the previous call to writematrix.