MANY THINGS (READ BELOW)
We fixed bugs that were in the previous version (dealing with the
function library stuffs).

We also allow solve_model to accept a directory where output files
are stored, so that whoever calls solve_model gets to decide where
they get their CSV/JSON files.

We also adjusted the code so that a CSV/JSON pair of files produced
during the same call to solve_model will have the same timestamp.

We also changed heatmap_demo to demonstrate these new features. Now,
every call to heatmap_demo will create a directory for its output based
on the date and time that heatmap_demo was called.