Finished writing to CSV file...
Finally got computation of the x and y components of the normal force
working, so those were added to the 'sol' matrix that solve_model
returns, and those numbers get written to the csv file.

Note that right now, there are some sets of input parameters that will
cause the code to crash/error/warning. It's unclear whether this is due
to a bug in the code, or if this is because we just have a complex
physical system, and we can't change inputs willy-nilly while expecting
a clean output. I suspect it is the latter though, so I think the code
works fine and the physics is just complicated.