Fixed t_L_guess bug/shortcut...
solve_model relies on calling ODE45 to solve the ODE to get the motion
of the LAMSA system. Ideally, we'd like to just numerically solve this
ODE until the system unlatches, and then run the code for the next
phases of motion. However, there's not an easy way to tell matlab
'numerically integrate until the system is doing this, and I have no
idea how long that will take.' What we do have is 'numerically integrate
until the system is doing this, or until you hit t_final, whichever
comes first'. So, the solution to this is to simply put the numerical
integration in a while-loop, and tell matlab to keep increasing t_final
until the numerical integration runs long enough to get stopped early by
the event (unlatching_end).