Fixed 'Sticking latch bug'...
Previously, solve_model had a bug where it would assume that the latch
gets stuck, and is unable to unlatch, if the maximum force exerted by
the unlatching motor is less than the frictional force. This is OK if
the latch starts with 0 velocity, but it's possible that if the latch
starts with a fast enough initial velocity, the system can still
unlatch.

To fix this, we have added a condition to unlatching_end, so that while
doing the numerical integration, an event is triggered if the latch has
velocity and acceleration close to 0. If this happens, we throw an error
in unlatching_end, and catch that error in solve_model. When this error
is caught, we simply return with sol=[initial conditions], and
transition_times = [infinity, infinity].

We also had to change get_metrics, because in the previous version, it
would throw an error if given a transition time of infinity.
Please enter the commit message for your changes. Lines starting