Added file-logging for latch getting stuck...
Previously, if the latch got stuck, solve_model would just return, and
not write to any files. However, it's probably a good idea to just write
the csv output (which is just initial conditions) and json input
data/metadata files anyway. So now the code does that.

To make this fix, we needed multiple places in the code to be able to
write to files. So, we decided to make a helper function to write that
info to the files, instead of copy-pasting code.