Added Temporary LaMSA Zone Plots
This is just a temporary testing file to see if we can create LaMSA zone plots using solve_direct_actuation. This is more or less a direct copy of heatmap_demo but with an added call  in the for loop to create a ratio in outvals instead of just a value. Also added transition_times values to solve_direct_actuation in order to be able to call it to get_metrics.