Began work for exploring parameter space...
So far, I've added code for springs and the motors that the
microrobotics folks are using. There's also some functions in there that
convert voltages to v_max and F_max, so we should be able to vary the
motor voltages relatively easily for when we loop over them later.

Currently, at 6V, there's no negative work possible. But we can try
changing the voltages, and we can also mess around with the motor
specs, which we're not entirely sure about...