Output Adjustments
Added a nargin statement that allows output directory to be optional. This will temporarily allow for runs to keep from outputting the csv and json files. Just leave out the final input to solve_model. However, in order to keep the plots that are generated, I've added in a statement that will save the figure produced and allowed input for a filename. Be sure to remember to name the file once the heatmap or LaMSA zone file stops running.