Fixed force issues in solve_model...
I found issues in solve_model that shouldn't affect anyone's physics, hopefully. These errors occur after the ODE45 calls, so the numerical solutions for x(t), xdot(t), y(t), ydot(t), should all be correct.

The errors have to do with how we calculated the various forces acting on the load and latch AFTER the calls to ODE45.

These correspond to the columns in 'sol' for the friction force, and to the 'unlatching motor force' column.

the unlatcihng motor force was doing a weird thing where the force discontinously jumped to the initial force after unlatching (now it just drops to 0, because we don't care about what happens to the latch after unlatching)

Furthermore, the friction forces were not being calculated correctly. Now they are.