Added plot_app stuff for git comparing!
previously, viewing changes on github to the app was difficult. This is because git doesn't like working with mlapp files, so you can't see changes to the app side-by-side in git.

Now, whenever the plot_app is run, the plot_app saves a file called plot_app.m which is identical to the 'code view' in the app, which allows