Fixed runway bug...
There was a weird edge case we found in the old code, which occurs when
x = runwaylength - radius
or
x = radius - runwaylength
This is a pretty rare edge case, but low-and-behold, we ran into it!

The issue is that we were multiplying 0*(1/0), which results in NaN. However, we realized we just wanted the function to evaluate to zero in this case. The fix was to do something more like 0*min(1/0, realmax).